+++ resolwe/flow/views/mixins.py
@@ -7,29 +7,7 @@ from rest_framework.decorators import list_route
 from rest_framework.generics import get_object_or_404
 from rest_framework.response import Response
 
-from resolwe.flow.models import DescriptorSchema
-from resolwe.permissions.shortcuts import get_objects_for_user
 from resolwe.permissions.utils import assign_contributor_permissions
-
-
-def get_descriptor_schames(query, user):
-    """Return filtered queryset od DescriptorSchama-s.
-
-    Queryset is filtered  by primary key (if ``query`` parameter is of
-    type ``int``` or slug (otherwise). Additionaly filter by ``VIEW``
-    permission is aplied for given user ``user``.
-    """
-    queryset_filter = {}
-
-    if isinstance(query, int):
-        queryset_filter['pk'] = query
-    else:
-        queryset_filter['slug'] = query
-
-    queryset = DescriptorSchema.objects.filter(**queryset_filter)
-    queryset = get_objects_for_user(user, 'view_descriptorschema', queryset)
-
-    return queryset
 
 
 class ResolweCreateModelMixin(mixins.CreateModelMixin):
@@ -44,23 +22,16 @@ class ResolweCreateModelMixin(mixins.CreateModelMixin):
 
     """
 
-    def create(self, request, *args, **kwargs):
-        """Create a resource."""
-        ds_query = request.data.get('descriptor_schema', None)
-        if ds_query:
-            ds_query = get_descriptor_schames(ds_query, request.user)
-            try:
-                request.data['descriptor_schema'] = ds_query.latest().pk
-            except DescriptorSchema.DoesNotExist:
-                return Response(
-                    {'descriptor_schema': [
-                        'Invalid descriptor_schema slug "{}" - object does not exist.'.format(ds_query)]},
-                    status=status.HTTP_400_BAD_REQUEST)
-
+    def define_contributor(self, request):
+        """Define contibutor by adding it to request.data."""
         if request.user.is_anonymous:
             request.data['contributor'] = get_anonymous_user().pk
         else:
             request.data['contributor'] = request.user.pk
+
+    def create(self, request, *args, **kwargs):
+        """Create a resource."""
+        self.define_contributor(request)
 
         try:
             return super().create(request, *args, **kwargs)
@@ -81,26 +52,10 @@ class ResolweUpdateModelMixin(mixins.UpdateModelMixin):
     """Mixin to support updating `Resolwe` models.
 
     Extends `django_rest_framework`'s class `UpdateModelMixin` with:
-
-      * translate `descriptor_schema` field from DescriptorSchema's
-        slug to its id and return 400 error Response if it doesn't
-        exists
-
     """
 
     def update(self, request, *args, **kwargs):
         """Update a resource."""
-        ds_query = request.data.get('descriptor_schema', None)
-        if ds_query:
-            ds_query = get_descriptor_schames(ds_query, request.user)
-            try:
-                request.data['descriptor_schema'] = ds_query.latest().pk
-            except DescriptorSchema.DoesNotExist:
-                return Response(
-                    {'descriptor_schema': [
-                        'Invalid descriptor_schema slug "{}" - object does not exist.'.format(ds_query)]},
-                    status=status.HTTP_400_BAD_REQUEST)
-
         # NOTE: Use the original method instead when support for locking is added:
         #       https://github.com/encode/django-rest-framework/issues/4675
         # return super().update(request, *args, **kwargs)